requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
langchain==0.0.340
openai==0.28.1
python-dotenv==1.0.1
faiss-cpu==1.7.4
tiktoken==0.5.1
